BazingaBoard Update on 02/17/16
Hello

The first Meteor app/project that I am going to be working on is going
to be BazingaBoard, which is going to be like the Leaderboard app that
I made in “Building Your First Meteor Application”. I will continue to
update it as I keep on learning/refining my Meteor skills!

Thank You for checking out BazingaBoard and God Bless!

—Jonathan Bruce